requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
SQLAlchemy==2.0.38
psycopg2-binary==2.9.10
pydantic-settings==2.8.1
python-jose==3.3.0
itsdangerous==2.2.0
jinja2==3.1.5
python-multipart==0.0.20
